Set access rights on your project

This function is used to record the permission levels for the deployment. The validity of the given values depend on the deployment type and the users/groups that have been registered there. It will only be checked at deployment time. In general, permission is understood as "viewing right", since editing all should happen through the git repository anyway.

use_access(users = NULL, groups = NULL, global = NULL)

Arguments

users

A character vector of user names that should have access to the deployment

groups

A character vector of group names that should have access to deployment

global

A global setting for permissions. For Posit Connect this can be "all" or "logged-in".

Value

The function is called for its side effects.