Provide additional settings for deployment

These functions will record additional deployment specific settings for the project, on top of the settings that are otherwise recorded in other DESCRIPTION fields.

Usage

use_connect_settings(...)

use_r_settings(...)

use_gha_settings(...)

Arguments

...

key-value pairs of settings. See details for allowed names. Set a key to NULL to remove it and thus use the default.

Value

This function is called for it's sideeffect

Details

Settings are specific to a given deployment target. Currently only Posit Connect is supported.

use_connect_settings()

Possible settings for Posit Connect are:

locked

locked_message

connection_timeout

read_timeout

init_timeout

idle_timeout

max_processes

min_processes

max_conns_per_process

load_factor

cpu_request

cpu_limit

memory_request

memory_limit

amd_gpu_limit

nvidia_gpu_limit

default_image_name

default_r_environment_management

default_py_environment_management

Please consult the API documentation for your Connect instance for description and allowed values for the various settings.