Execute a deployed project on a schedule

For projects that does not run continuously, such as APIs or apps, it may make sense to execute them periodically rather than only once when deployed or manually when you trigger it. This function records an intended execution schedule in the DESCRIPTION file that will be set during deployment. Do note that not all deployment environments support all scheduling settings. If it is deducible what type of deployment you are using the provided schedule will be tested before recording it.

Usage

use_schedule(
  minute = NULL,
  hour = NULL,
  day = NULL,
  weekday = NULL,
  week = NULL,
  month = NULL,
  year = NULL
)

at(x, to = NULL, step = NULL)

interval(x)

Arguments

minute, hour, day, weekday, week, month, year

Scheduling times. See Datails on how to set them

x

An interval or time location

to

The end of the schedule range

step

The step size of the schedule range